# docs/swagger.yaml
basePath: /api
definitions:
...
  global.MODULE:
    properties:
      active:
        type: string
      comment:
        type: string
      create_date:
        type: string
      create_user:
        type: integer
      icon:
        type: string
      irl:
        type: string
      module_id:
        type: integer
      module_name:
        type: string
      module_seq:
        type: string
      parent_module:
        type: string
      update_date:
        type: string
      update_user:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a Api Documentation for SIPIL 2022 with Golang Backend, Gin
    Gonic Framework, GORM MySQL, and JWT Authentication
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API DOCUMENTATION - API FOR SIPIL 2022
  version: "1.0"
paths:
...
  /employee:
    get:
      consumes:
      - application/json
      description: API untuk mengambil semua employee yang terdapat dalam database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Mengambil Semua Employees
      tags:
      - Employees
    post:
      consumes:
      - application/json
      description: API untuk membuat employee baru.
      parameters:
      - description: Employees Data
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/global.MODULE'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Membuat employee baru.
      tags:
      - Employees
  /employee/{id}:
    delete:
      consumes:
      - application/json
      description: API untuk delete employee.
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Delete employee.
      tags:
      - Employees
    get:
      consumes:
      - application/json
      description: API untuk mencari employee by ID yang terdapat dalam database.
      parameters:
      - description: Employees ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Mengambil Employees by ID
      tags:
      - Employees
    put:
      consumes:
      - application/json
      description: API untuk update employee.
      parameters:
      - description: Employees Data
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/global.MODULE'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Update employee.
      tags:
      - Employees
...
  /location:
    get:
      consumes:
      - application/json
      description: API untuk mengambil semua locationyang terdapat dalam database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Mengambil Semua Locations
      tags:
      - Locations
    post:
      consumes:
      - application/json
      description: API untuk membuat locationbaru.
      parameters:
      - description: Locations Data
        in: body
        name: location
        required: true
        schema:
          $ref: '#/definitions/global.MODULE'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Membuat locationbaru.
      tags:
      - Locations
  /location/{id}:
    delete:
      consumes:
      - application/json
      description: API untuk delete location by ID
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Delete location
      tags:
      - Locations
    get:
      consumes:
      - application/json
      description: API untuk mencari locationby ID yang terdapat dalam database.
      parameters:
      - description: Locations ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Mengambil Locations by ID
      tags:
      - Locations
    put:
      consumes:
      - application/json
      description: API untuk update location
      parameters:
      - description: Locations Data
        in: body
        name: location
        required: true
        schema:
          $ref: '#/definitions/global.MODULE'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Update location
      tags:
      - Locations
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"